requirements.txt
hydra-core==1.3.2
numpy
wandb
easydict
transformers
opencv-python
einops
thop
robosuite==1.4.0
bddl
future
matplotlib
cloudpickle
gym
numexpr
h5py
torchsde
torchdiffeq
imageio[ffmpeg]
open3d
termcolor